+++ src/defaults.js
@@ -68,7 +68,14 @@ export default {
   mimeType: 'auto',
 
   /**
-   * PNG files over this value (5 MB by default) will be converted to JPEGs.
+   * Files whose file type is included in this list,
+   * and whose file size exceeds the `convertSize` value will be converted to JPEGs.
+   * @type {string｜Array}
+   */
+  convertTypes: ['image/png'],
+
+  /**
+   * PNG files over this size (5 MB by default) will be converted to JPEGs.
    * To disable this, just set the value to `Infinity`.
    * @type {number}
    */